requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.1.1
waitress==3.0.2
APScheduler==3.10.4
python-dateutil==2.8.2